Moved tsx files into their own folder
