+++ leaguewordle-frontend/package.json
@@ -14,7 +14,6 @@
     "react-router-dom": "^5.3.3",
     "react-scripts": "^5.0.1",
     "react-sliding-side-panel": "^2.0.3",
-    "react-spring": "^9.5.3",
     "react-transition-group": "^4.4.2",
     "web-vitals": "^2.1.4"
   },
